+++ internal/internal_worker_heartbeat_metrics.go
@@ -100,7 +100,7 @@ func (h *heartbeatMetricsHandler) Gauge(name string) metrics.Gauge {
 				value:      h.getOrCreate(name + ":" + h.workerType),
 			}
 		}
-	case metrics.NumPoller:
+	case metrics.NumPoller, metrics.PollerTarget:
 		if h.pollerType != "" {
 			return &capturingGauge{
 				underlying: underlying,
@@ -219,21 +219,25 @@ func (h *heartbeatMetricsHandler) PopulateHeartbeat(hb *workerpb.WorkerHeartbeat
 
 	hb.WorkflowPollerInfo = buildPollerInfo(
 		int32(h.get(metrics.NumPoller+":"+metrics.PollerTypeWorkflowTask)),
+		int32(h.get(metrics.PollerTarget+":"+metrics.PollerTypeWorkflowTask)),
 		opts.pollTimeTracker.getLastPollTime(metrics.PollerTypeWorkflowTask),
 		opts.workflowPollerBehavior,
 	)
 	hb.WorkflowStickyPollerInfo = buildPollerInfo(
 		int32(h.get(metrics.NumPoller+":"+metrics.PollerTypeWorkflowStickyTask)),
+		int32(h.get(metrics.PollerTarget+":"+metrics.PollerTypeWorkflowStickyTask)),
 		opts.pollTimeTracker.getLastPollTime(metrics.PollerTypeWorkflowStickyTask),
 		opts.workflowPollerBehavior,
 	)
 	hb.ActivityPollerInfo = buildPollerInfo(
 		int32(h.get(metrics.NumPoller+":"+metrics.PollerTypeActivityTask)),
+		int32(h.get(metrics.PollerTarget+":"+metrics.PollerTypeActivityTask)),
 		opts.pollTimeTracker.getLastPollTime(metrics.PollerTypeActivityTask),
 		opts.activityPollerBehavior,
 	)
 	hb.NexusPollerInfo = buildPollerInfo(
 		int32(h.get(metrics.NumPoller+":"+metrics.PollerTypeNexusTask)),
+		int32(h.get(metrics.PollerTarget+":"+metrics.PollerTypeNexusTask)),
 		opts.pollTimeTracker.getLastPollTime(metrics.PollerTypeNexusTask),
 		opts.nexusPollerBehavior,
 	)
@@ -269,7 +273,12 @@ func buildSlotsInfo(
 	}
 }
 
-func buildPollerInfo(currentPollers int32, lastSuccessfulPollTime time.Time, pollerBehavior PollerBehavior) *workerpb.WorkerPollerInfo {
+func buildPollerInfo(
+	currentPollers int32,
+	targetPollers int32,
+	lastSuccessfulPollTime time.Time,
+	pollerBehavior PollerBehavior,
+) *workerpb.WorkerPollerInfo {
 	var isAutoscaling bool
 	switch pollerBehavior.(type) {
 	case *pollerBehaviorAutoscaling:
@@ -282,6 +291,7 @@ func buildPollerInfo(currentPollers int32, lastSuccessfulPollTime time.Time, pol
 
 	return &workerpb.WorkerPollerInfo{
 		CurrentPollers:         currentPollers,
+		TargetPollers:          targetPollers,
 		LastSuccessfulPollTime: pollTime,
 		IsAutoscaling:          isAutoscaling,
 	}